backend/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
requests==2.26.0
beautifulsoup4==4.10.0
tensorflow==2.7.0
tensorflow-hub==0.12.0
APScheduler==3.7.0
gunicorn==20.1.0
pandas==1.3.3
python-dotenv==0.19.1
Werkzeug==2.0.2
typing_extensions==4.3.0
SQLAlchemy==1.4.31
Flask-CORS==3.0.10

